requirements.txt
accelerate==1.0.1
addict==2.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
attrs==24.2.0
Authlib==1.3.2
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.0
cryptography==43.0.1
cupy==13.3.0
cycler==0.12.1
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
Deprecated==1.2.14
dimod==0.12.17
diskcache==5.6.3
dwave-cloud-client==0.13.1
dwave-neal==0.6.0
dwave-optimization==0.4.0
dwave-preprocessing==0.6.7
dwave-samplers==1.4.0
dwave-system==1.27.0
dwave_networkx==0.8.15
exceptiongroup==1.2.2
executing==2.0.1
fasteners==0.19
fastjsonschema==2.20.0
fastrlock==0.8.2
filelock==3.13.1
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.2.0
ftfy==6.2.3
homebase==1.0.1
huggingface-hub==0.24.6
hydra-core==1.3.2
idna==3.8
importlib_metadata==8.4.0
iopath==0.1.10
ipython==8.26.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
lightning-utilities==0.11.6
llvmlite==0.40.1
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
minorminer==0.2.15
mpmath==1.3.0
munch==4.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.57.1
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==12.535.161
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
nvitop==1.3.2
omegaconf==2.3.0
open3d==0.18.0
open_clip_torch==2.26.1
opencv-python==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
plotly==5.24.0
plucky==0.4.3
plyfile==1.1
pointnet2 @ file:///root/3dllm/minhlnh/FreeVocab-3DIS/util3d/pointnet2/dist/pointnet2-0.0.0-cp310-cp310-linux_x86_64.whl#sha256=0b7e4b0f068aa3e0a1d7041e6052a6b1ab1ba026041aa9e8449bb11ad8b7e8d7
portalocker==2.10.1
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
PuLP==2.9.0
pure_eval==0.2.3
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
pyquaternion==0.9.9
pyqubo==1.4.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
pyviz3d==0.3.5
PyYAML==6.0.2
qubobrute==0.1.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
retrying==1.3.4
rpds-py==0.20.0
safetensors==0.4.5
-e git+https://github.com/PhucNDA/FreeVocab-3DIS.git@54cdf1d3cb60df03fb64924f7863ca2f6c3e22eb#egg=SAM_2&subdirectory=segmenter2d/segment-anything-2
scikit-learn==1.5.1
scipy==1.14.1
segment-anything-hq==0.3
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
shrdr @ file:///root/3dllm/minhlnh/FreeVocab-3DIS/shrdr
six==1.16.0
stack-data==0.6.3
sympy==1.12
tenacity==9.0.0
termcolor==2.4.0
threadpoolctl==3.5.0
timm==1.0.9
tokenizers==0.13.3
torch==2.3.1+cu118
torch_scatter @ file:///root/3dllm/minhlnh/torch_scatter-2.1.2%2Bpt23cu118-cp310-cp310-linux_x86_64.whl#sha256=d05f8a4f2a23169b6e7ad62f67d05d470266dafca0a7cdbb9aad69c3f5d88e11
torchaudio==2.3.1+cu118
torchmetrics==1.4.1
torchvision==0.18.1+cu118
tqdm==4.66.5
traitlets==5.14.3
transformers==4.33.1
triton==2.3.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==1.16.0
zipp==3.20.1
